+++ cli_tool_audit/call_tools.py
@@ -10,7 +10,9 @@ Possible things that could happen
 - found, has version but cli version != package version
 """
 
+import datetime
 import logging
+import os
 import subprocess  # nosec
 from dataclasses import dataclass
 from typing import Optional
@@ -18,7 +20,6 @@ from typing import Optional
 # pylint: disable=no-name-in-module
 from whichcraft import which
 
-from cli_tool_audit.config_reader import read_config
 from cli_tool_audit.known_swtiches import KNOWN_SWITCHES
 
 logger = logging.getLogger(__name__)
@@ -33,6 +34,20 @@ class ToolAvailabilityResult:
     is_available: bool
     is_broken: bool
     version: Optional[str]
+    last_modified: Optional[datetime.datetime]
+
+
+def get_command_last_modified_date(tool_name: str) -> Optional[datetime.datetime]:
+    # Find the path of the command's executable
+    result = which(tool_name)
+    if result is None:
+        return None
+
+    executable_path = result
+
+    # Get the last modified time of the executable
+    last_modified_timestamp = os.path.getmtime(executable_path)
+    return datetime.datetime.fromtimestamp(last_modified_timestamp)
 
 
 def check_tool_availability(
@@ -52,13 +67,13 @@ def check_tool_availability(
     # Check if the tool is in the system's PATH
     is_broken = True
 
-    # This is wrong?
-    if not which(tool_name):
+    last_modified = get_command_last_modified_date(tool_name)
+    if not last_modified:
         logger.warning(f"{tool_name} is not on path.")
-        return ToolAvailabilityResult(False, True, None)
+        return ToolAvailabilityResult(False, True, None, last_modified)
     if only_check_existence:
         logger.debug(f"{tool_name} exists, but not checking for version.")
-        return ToolAvailabilityResult(True, False, None)
+        return ToolAvailabilityResult(True, False, None, last_modified)
 
     if version_switch is None or version_switch == "--version":
         # override default.
@@ -87,25 +102,26 @@ def check_tool_availability(
         logger.error(f"{tool_name} stdout: {exception.stdout}")
     except FileNotFoundError:
         logger.error(f"{tool_name} is not on path.")
-        return ToolAvailabilityResult(False, True, None)
+        return ToolAvailabilityResult(False, True, None, last_modified)
 
-    return ToolAvailabilityResult(True, is_broken, version)
+    return ToolAvailabilityResult(True, is_broken, version, last_modified)
 
 
 if __name__ == "__main__":
-    check_tool_availability("isort")
-
-    def run() -> None:
-        """Example"""
-        # Example usage
-        file_path = "../pyproject.toml"
-        cli_tools = read_config(file_path)
-
-        for tool, config in cli_tools.items():
-            result = check_tool_availability(tool, config.get("version_switch", "--version"))
-            print(
-                f"{tool}: {'Available' if result.is_available else 'Not Available'}"
-                f" - Version:\n{result.version if result.version else 'N/A'}"
-            )
-
-    run()
+    print(get_command_last_modified_date("asdfpipx"))
+    # check_tool_availability("isort")
+    #
+    # def run() -> None:
+    #     """Example"""
+    #     # Example usage
+    #     file_path = "../pyproject.toml"
+    #     cli_tools = read_config(file_path)
+    #
+    #     for tool, config in cli_tools.items():
+    #         result = check_tool_availability(tool, config.version_switch or "--version")
+    #         print(
+    #             f"{tool}: {'Available' if result.is_available else 'Not Available'}"
+    #             f" - Version:\n{result.version if result.version else 'N/A'}"
+    #         )
+    #
+    # run()