+++ CHANGES.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2024-01-14
+
+### Added
+- Support for if_os, snapshot
+- Config manager and possibility to do config via cli
+- Freeze command
+- Last modified
+
+### Fixed
+- Tested with tox and got passing on 3.9-3.12
+- Possibly fixed dependencies.
+
 ## [1.0.3] - 2024-01-14
 
 ### Added